Refactor (#7)
* yolov5 submodule

black + isort

* changes in dataset link parser

* better scripts

* script to download all the datasets

* black + isort

* download_datasets.sh

* docker file + doc

* typos

* minor changes

* train yolov7 inside docker

* train for yolov5/7 + docker + doc

* minor changes

* datasets links (with 640)

* glip runs but it fails

* minor changes

* fixed issue with yolov5

* readme for init submodules

* train with model before reparam

* update dataset list to included filter extremely underrepresented classes

* update gauge dataset

* results

* script for table

* minor changes

* conflicts

+++ GLIP-benchmark/train.sh
@@ -35,19 +35,22 @@ for dataset in $(ls $datasets)
 do
     dataset=$datasets/$dataset
     echo "Training on $dataset"
-    ln -s $dataset/test ./DATASET/coco/val2017
-    ln -s $dataset/test/_annotations.coco.json ./DATASET/coco/annotations/instances_val2017.json
+    if [ ! -d "$dataset/results" ] ;
+    then
+        ln -s $dataset/test ./DATASET/coco/val2017
+        ln -s $dataset/test/_annotations.coco.json ./DATASET/coco/annotations/instances_val2017.json
 
-    PYTHONPATH=""
-    # interesting, we need to append this module to the PYTHONPATH
-    export PYTHONPATH="${PYTHONPATH}:${PWD}/"  && python tools/test_grounding_net.py --config-file ../glip_Swin_T_O365_GoldG.yaml --weight $dir/glip_tiny_model_o365_goldg_cc_sbu.pth \
-        TEST.IMS_PER_BATCH 1 \
-        MODEL.DYHEAD.SCORE_AGG "MEAN" \
-        TEST.EVAL_TASK detection \
-        MODEL.DYHEAD.FUSE_CONFIG.MLM_LOSS False \
-        OUTPUT_DIR $dataset/results
-    rm ./DATASET/coco/val2017
-    rm ./DATASET/coco/annotations/instances_val2017.json
+        PYTHONPATH=""
+        # interesting, we need to append this module to the PYTHONPATH
+        export PYTHONPATH="${PYTHONPATH}:${PWD}/"  && python tools/test_grounding_net.py --config-file ../glip_Swin_T_O365_GoldG.yaml --weight $dir/glip_tiny_model_o365_goldg_cc_sbu.pth \
+            TEST.IMS_PER_BATCH 1 \
+            MODEL.DYHEAD.SCORE_AGG "MEAN" \
+            TEST.EVAL_TASK detection \
+            MODEL.DYHEAD.FUSE_CONFIG.MLM_LOSS False \
+            OUTPUT_DIR $dataset/results
+        rm ./DATASET/coco/val2017
+        rm ./DATASET/coco/annotations/instances_val2017.json
+    fi
 done
 
 echo "Done evaluating all the datasets with GLIP!"
